# openapi.yaml
openapi: 3.1.0
info:
  title: WHOOP Health Coach API
  description: Complete API for WHOOP health data
  version: 2.0.0
servers:
  - url: https://whoop-gpts-backend.onrender.com
paths:
...
  /whoop/sleep/history:
    get:
      operationId: getSleepHistory
      summary: Get sleep history for the past days
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Success
...
  /whoop/recovery/history:
    get:
      operationId: getRecoveryHistory
      summary: Get recovery history for the past days
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Success
...
  /whoop/workout/history:
    get:
      operationId: getWorkoutHistory
      summary: Get workout history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Success
...